+++ src/main/java/de/codeinfection/quickwango/ItemRepair/RepairBlocks/NormalRepair.java
@@ -1,15 +1,13 @@
 package de.codeinfection.quickwango.ItemRepair.RepairBlocks;
 
-import de.codeinfection.quickwango.ItemRepair.Item;
 import de.codeinfection.quickwango.ItemRepair.ItemRepair;
 import de.codeinfection.quickwango.ItemRepair.RepairBlock;
 import de.codeinfection.quickwango.ItemRepair.RepairRequest;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import org.bukkit.ChatColor;
+import static de.codeinfection.quickwango.Translation.Translator.t;
+import java.util.Map;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
+import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 
 /**
@@ -22,7 +20,7 @@ public class NormalRepair extends RepairBlock
 
     public NormalRepair(Material material)
     {
-        super(ItemRepair.getInstance(), "complete", material);
+        super(ItemRepair.getInstance(), t("repair"), material);
     }
 
     public NormalRepair(int blockId)
@@ -36,43 +34,24 @@ public class NormalRepair extends RepairBlock
     }
 
     @Override
-    public RepairRequest requestRepair(Player player)
+    public RepairRequest requestRepair(Inventory inventory)
     {
-        if (hasPermission(player))
+        final Player player = (Player)inventory.getHolder();
+        Map<Integer, ItemStack> items = getRepairableItems(inventory);
+        if (items.size() > 0)
         {
-            double price;
-            ArrayList<ItemStack> allItems = new ArrayList<ItemStack>();
-            Collections.addAll(allItems, player.getInventory().getArmorContents());
-            Collections.addAll(allItems, player.getInventory().getContents());
-            List<ItemStack> items = new ArrayList<ItemStack>();
-            ItemStack itemInHand = player.getItemInHand();
+            double price = calculatePrice(items.values());
 
-            if (itemInHand != null && Item.getByMaterial(itemInHand.getType()) != null)
-            {
-                for (ItemStack itemStack : allItems)
-                {
-                    if (itemStack != null && Item.getByMaterial(itemStack.getType()) != null && itemStack.getDurability() > 0)
-                    {
-                        items.add(itemStack);
-                    }
-                }
+            player.sendMessage(t("headline"));
+            player.sendMessage(t("rightClickAgain"));
+            player.sendMessage(t("repairWouldCost", getEconomy().format(price)));
+            player.sendMessage(t("youCurrentlyHave", getEconomy().format(getEconomy().getBalance(player.getName()))));
 
-                if (items.size() > 0)
-                {
-                    price = calculatePrice(items);
-
-                    player.sendMessage(ChatColor.GREEN + "[" + ChatColor.DARK_RED + "ItemRepair" + ChatColor.GREEN + "]");
-                    player.sendMessage(ChatColor.AQUA + "Rightclick" + ChatColor.WHITE + " again to repair all your damaged items.");
-                    player.sendMessage("The repair would cost " + ChatColor.AQUA + getEconomy().format(price) + ChatColor.WHITE + ".");
-                    player.sendMessage("You have currently " + ChatColor.AQUA + getEconomy().format(getEconomy().getBalance(player.getName())));
-
-                    return new RepairRequest(this, player, items, price);
-                }
-                else
-                {
-                    player.sendMessage(ChatColor.RED + "You don't have any items to repair!");
-                }
-            }
+            return new RepairRequest(this, inventory, items, price);
+        }
+        else
+        {
+            player.sendMessage(t("noItems"));
         }
         return null;
     }
@@ -80,30 +59,24 @@ public class NormalRepair extends RepairBlock
     @Override
     public void repair(RepairRequest request)
     {
-        double price = request.getPrice();
-        Player player = request.getPlayer();
+        final double price = request.getPrice();
+        final Player player = (Player)request.getInventory().getHolder();
 
         if (getEconomy().getBalance(player.getName()) >= price)
         {
-            List<ItemStack> items = request.getItems();
-            ItemStack itemInHand = player.getItemInHand();
-            if (Item.getByMaterial(itemInHand.getType()) != null)
-            {
-                items.add(itemInHand);
-            }
             if (getEconomy().withdrawPlayer(player.getName(), price).transactionSuccess())
             {
-                repairItems(items);
-                player.sendMessage(ChatColor.GREEN + "Your items have been repaired for " + ChatColor.AQUA + getEconomy().format(price));
+                repairItems(request);
+                player.sendMessage(t("itemsRepaired", getEconomy().format(price)));
             }
             else
             {
-                player.sendMessage(ChatColor.RED + "Something went wrong, report this failure to your administrator!");
+                player.sendMessage(t("somethingWentWrong"));
             }
         }
         else
         {
-            player.sendMessage(ChatColor.RED + "You don't have enough money to repair all your items!");
+            player.sendMessage(t("notEnoughMoney"));
         }
     }
 }